hw3

+++ services/contacts.js
@@ -1,35 +1,37 @@
 const { Contact } = require("../db/contactModel");
 const { ParameterError } = require("../helpers/errors");
 
-const listContacts = async () => {
-  const contacts = await Contact.find();
+const listContacts = async (owner,page,limit, favorite=null) => {
+  const skip = page <= 0 ? 0 : (page-1) * limit;
+  const filter = favorite ? { owner, favorite } : { owner };
+  const contacts = await Contact.find(filter, {__v: 0, owner: 0}).skip(skip).limit(limit)
   return contacts;
 };
 
-const getContactById = async (id) => {
-  const contact = await Contact.findOne({ _id: id });
+const getContactById = async (id, owner) => {
+  const contact = await Contact.findOne({ _id: id, owner }, {__v: 0, owner: 0});
   if (!contact) {
     throw new ParameterError(`Not found!`);
   }
   return contact;
 };
 
-const removeContact = async (id) => {
-  const removedContact = await Contact.findOneAndRemove({ _id: id });
+const removeContact = async (id, owner) => {
+  const removedContact = await Contact.findOneAndRemove({ _id: id, owner });
   return removedContact;
 };
 
-const addContact = async (name, phone, email, favorite = false) => {
-  const newContact = new Contact({ name, phone, email, favorite });
+const addContact = async (name, phone, email, favorite = false, owner) => {
+  const newContact = new Contact({ name, phone, email, favorite, owner });
   await newContact.save();
   return newContact;
 };
 
-const updateContact = async (id, { name, phone, email, favorite }) => {
+const updateContact = async (id, { name, phone, email, favorite }, owner) => {
   const contact = await Contact.findOneAndUpdate(
-    { _id: id },
+    { _id: id, owner },
     { $set: { name, phone, email, favorite } },
-    {new: true}
+    {new: true},
   );
   if (!contact) {
     throw new ParameterError(`Not found!`);
@@ -37,9 +39,9 @@ const updateContact = async (id, { name, phone, email, favorite }) => {
   return contact;
 };
 
-const updateStatusContact = async (id, {favorite}) => {
+const updateStatusContact = async (id, {favorite}, owner) => {
 const contact = await Contact.findOneAndUpdate(
-    { _id: id },
+    { _id: id,owner },
     { $set: { favorite } },
     {new: true}
   );
